backend/requirements.txt
django
djangorestframework
psycopg2-binary
pytest
pytest-django
pytest-cov